implemented a more general Texture class
Implemented a more general Texture class, which allows a mesh to process all textures no matter how they were generated. Before this class was implemented a texture for a mesh had to be loaded from an image but now it is possible to use textures that were attached to a framebuffer, for example. The TextureImage class inherits the Texture class.