added multisampling support for textures and renderbuffers
This update implemented basic multisampling support for textures and renderbuffers. To use multisampling the classes "MultisampleTexture" and "MultisampleRenderBuffer" should be used. A new TextureType "Multisample" was added, which is used by these classes. The format of that TextureType is GL_RGBA16F (float for HDR). RGBA16F can now also be used for normal textures.