implemented basic camera
Implemented a simple camera. The class "Camera" can now be used to simulate a simple camera and to build own camera models based on this one. Please note that this class only provides support for a Camera, which looks on a target. A fps-camera will be implemented later based on this class. This commit also adds the class "ScreenModel", which inherits from "Model" and has a mesh preloaded consisting out of two, screen-filling triangles (it can be used to render a texture to the whole screen, e.g. from a custom framebuffer).