reworked UniformBuffer
The UniformBuffer class now manages the memory allocated for the buffer by its self. The class will allocate more memory as soon as a new element is added, which does not fit into the current allocated buffer (capacity). Because of this, the size of the UniformBuffer should be given in the constructor (if possible) in bytes. The size of one UniformType can be queried with UniformTypes::getSize(UniformType type). This class assumes the buffer uses the std140 layout. This fact can also be used to calculate the overall size.