implemented simple diffuse IBL
Implemented simple, diffuse, image based lighting. A 'CubeMapArray' class was added, which is used for IBL. This class also can convert a equirectengular 'HDRTexture' to a "normal" cubemap:
otg::CubeMapArray cubemap(512, 512);
cubemap.fromEquirectengular(hdrTexture);

Currently this implementation only supports IBL with a static cubemap (e.g. skybox) that is convoluted before the rendering loop was started. How that cubemap was filled with data is up to the user. The new 'SkyBoxLightProbe' class can be used for basic diffuse IBL. New shaders were added for IBL (e.g. for the conversion of a equirectengular texture to a cubemap).
+bugfixes, refactoring