implemented SkyBoxReflectionProbe
Implemented simple, specular image based lighting. To use specular IBL you can use the 'SkyBoxReflectionProbe' class. This class should be used for IBL through a skybox (i.e. a immutable 'CubeMapArray'). This probe will bake two textures: brdfIntegrationMap and brdfTexture. Both can be bound to a texture-unit with SkyBoxReflectionProbe::bindPrefilterToUnit and SkyBoxReflectionProbe::bindLookUpTexToUnit. The prefilter-map is a 'CubeMapArray' and the lookup-texture a normal 2D-texture.
+bugfix: model-/worldmatrix of models was not correctly initilized
*refactor: SkyBoxLightProbe