reworked material and implemented basic heightmap class
The material of a mesh is now responsible for the corresponding textures and not the mesh itself. A specific 'Material' class was created to support this behavior. This class replaces the structure that was used before.
The 'Heightmap<T>' class can now be used to store a 2d grid of values (of type T).